chore : 소스 통합 후 임시로 만든 클래스들 삭제
