+++ src/main/java/study/toy/everythingshop/entity/mariaDB/Product.java
@@ -1,17 +1,12 @@
 package study.toy.everythingshop.entity.mariaDB;
 
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import lombok.Getter;
+import study.toy.everythingshop.enums.ProductStatus;
 
-@Data
-@Builder //상속받는 클래스가 없으므로 SuperBuilder가 아닌 builder사용
-@NoArgsConstructor
-@AllArgsConstructor
+@Getter
 public class Product {
     private Integer productNum;         //상품번호
-    private Integer userNum;            //사용자번호(판매자)
+    private User user;                  //사용자
     private String productNm;           //상품명
     private Integer registerQuantity;   //등록수량
     private Integer remainQuantity;     //남은수량
@@ -20,6 +15,7 @@ public class Product {
     private String postYn;              //게시여부(Y, N)
     private String registerDt;          //등록일자
     private String changeDt;            //수정일자
+    private ProductViews productViews;
 
     public void getOrdered(Integer orderedQty){
         this.remainQuantity = remainQuantity - orderedQty;
@@ -27,7 +23,20 @@ public class Product {
     }
     public void updateStts(){
         if(remainQuantity == 0){
-            productStatusCd = "03" ;
+            productStatusCd = ProductStatus.SOLD_OUT.getCode();
         }
+    }
+    public Integer getViews() {
+        return productViews.getViews();
+    }
+    public void increaseView() {
+        this.productViews.increaseView();
+    }
+    public Integer getDiscountPrice(int discountRate) {
+        int discountPrice = (int) Math.round(this.productPrice * (discountRate / 100.0)) ;
+        return (int) Math.ceil(discountPrice / 10.0) * 10; //1원단위 반올림
+    }
+    public Integer getSalesQuantity() {
+        return this.registerQuantity - this.remainQuantity;
     }
 }
