+++ src/main/java/study/toy/everythingshop/entity/mariaDB/ProductViews.java
@@ -1,17 +1,15 @@
 package study.toy.everythingshop.entity.mariaDB;
 
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
+import lombok.Getter;
 
-@Data
-@Builder //상속받는 클래스가 없으므로 SuperBuilder가 아닌 builder사용
-@NoArgsConstructor
-@AllArgsConstructor
+@Getter
 public class ProductViews {
     private Integer productNum;         //상품번호
     private Integer views;              //조회수
     private String registerDt;          //등록일자
     private String changeDt;            //수정일자
+
+    public void increaseView() {
+        this.views += 1;
+    }
 }
