+++ src/main/java/study/toy/everythingshop/repository/ProductDAO.java
@@ -3,7 +3,6 @@ package study.toy.everythingshop.repository;
 import study.toy.everythingshop.dto.*;
 import study.toy.everythingshop.entity.h2.ProductMEntity;
 import study.toy.everythingshop.entity.mariaDB.Product;
-import study.toy.everythingshop.entity.mariaDB.ProductN;
 
 import java.util.List;
 
@@ -19,9 +18,9 @@ public interface ProductDAO {
 
     ProductDTO selectByProductNum(Integer productNum);
 
-    ProductN selectProductsWithViews(Integer productNum);
+    Product selectProductsWithViews(Integer productNum);
 
-    int updateProductViews(ProductN product);
+    int updateProductViews(Product product);
 
     int insertProduct(ProductRegisterDTO productRegisterDTO);
