+++ src/main/java/study/toy/everythingshop/repository/impl/ProductDAOImpl.java
@@ -6,7 +6,6 @@ import org.springframework.stereotype.Repository;
 import study.toy.everythingshop.dto.*;
 import study.toy.everythingshop.entity.h2.ProductMEntity;
 import study.toy.everythingshop.entity.mariaDB.Product;
-import study.toy.everythingshop.entity.mariaDB.ProductN;
 import study.toy.everythingshop.logTrace.Trace;
 import study.toy.everythingshop.repository.ProductDAO;
 
@@ -101,12 +100,12 @@ public class ProductDAOImpl implements ProductDAO {
     }
   
     @Override
-    public ProductN selectProductsWithViews(Integer productNum) {
+    public Product selectProductsWithViews(Integer productNum) {
         return sqlSession.selectOne("maria.ProductDAO.selectProductsWithViews", productNum);
     }
 
     @Override
-    public int updateProductViews(ProductN product) {
+    public int updateProductViews(Product product) {
         return sqlSession.update("maria.ProductDAO.updateProductViews", product);
     }
 }
