+++ src/main/java/study/toy/everythingshop/service/impl/ProductServiceImpl.java
@@ -12,7 +12,6 @@ import study.toy.everythingshop.dto.*;
 import study.toy.everythingshop.entity.mariaDB.OrderedProduct;
 import study.toy.everythingshop.entity.mariaDB.PointHistory;
 import study.toy.everythingshop.entity.mariaDB.Product;
-import study.toy.everythingshop.entity.mariaDB.ProductN;
 import study.toy.everythingshop.entity.mariaDB.User;
 import study.toy.everythingshop.enums.CommonCodeClassEnum;
 import study.toy.everythingshop.logTrace.Trace;
@@ -38,8 +37,6 @@ public class ProductServiceImpl implements ProductService {
     private final UserDAO userDAO;
     private final DiscountPolicyDAO discountPolicyDAO;
     private final CommonService commonService;
-
-    Product product = new Product();
 
     @Value("${default.recordCountPerPage}")
     private int defaultRecordCountPerPage;
@@ -85,7 +82,7 @@ public class ProductServiceImpl implements ProductService {
     @Override
     public ProductDTO findProductDetail(Integer productNum, boolean firstView, UserDetails userDetails) {
         //상품조회
-        ProductN product = productDAO.selectProductsWithViews(productNum);
+        Product product = productDAO.selectProductsWithViews(productNum);
 
         //조회수 증가
         increaseViews(product, firstView);
@@ -109,14 +106,14 @@ public class ProductServiceImpl implements ProductService {
                 .build();
     }
 
-    private void increaseViews(ProductN product, boolean firstView) {
+    private void increaseViews(Product product, boolean firstView) {
         if(firstView) {
             product.increaseView();//조회수 증가
             productDAO.updateProductViews(product);//조회수 업데이트
         }
     }
 
-    private int calculateDiscountPrice(ProductN product, UserDetails userDetails) {
+    private int calculateDiscountPrice(Product product, UserDetails userDetails) {
         if(userDetails != null) {
             User user = userDAO.selectUserById(userDetails.getUsername());
             int discountRate = userDAO.selectUserDiscountRate(user.getUserNum());
